move new components into own files
